E-Commercial_Redux_Example
This project is a of fetching data from an API using createAsyncThunk, and implementing "Add to Cart" and "Remove from Cart" functionalities.it includes features to display the total amount and total items in the cart. This platform is an excellent resource for both beginners and those looking to solidify their knowledge of Redux.